[wip] chore: Moves web3-uri/hotel to web3-uri/providers/json-hotel
