+++ test/wt-libs/data-model/web3-uri/providers/json-hotel.spec.js
@@ -1,11 +1,11 @@
 import { assert } from 'chai';
 import sinon from 'sinon';
-import testedDataModel from '../../../utils/data-model-definition';
-import jsonWallet from '../../../utils/test-wallet';
-import Web3UriDataModel from '../../../../src/data-model/web3-uri';
-import HotelDataProvider from '../../../../src/data-model/web3-uri/hotel';
+import testedDataModel from '../../../../utils/data-model-definition';
+import jsonWallet from '../../../../utils/test-wallet';
+import Web3UriDataModel from '../../../../../src/data-model/web3-uri';
+import JsonHotelProvider from '../../../../../src/data-model/web3-uri/providers/json-hotel';
 
-describe('WTLibs.data-model.web3-uri.hotel', () => {
+describe('WTLibs.data-model.web3-uri.providers.json-hotel', () => {
   const correctPassword = 'test123';
   let dataModel, indexDataProvider;
 
@@ -32,7 +32,7 @@ describe('WTLibs.data-model.web3-uri.hotel', () => {
 
     it('should throw when we want hotel from a bad address', async () => {
       try {
-        const hotelProvider = await HotelDataProvider.createInstance(dataModel.web3Utils, dataModel.web3Contracts, await indexDataProvider.__getDeployedIndex(), '0x96eA4BbF71FEa3c9411C1Cefc555E9d7189695fA');
+        const hotelProvider = await JsonHotelProvider.createInstance(dataModel.web3Utils, dataModel.web3Contracts, await indexDataProvider.__getDeployedIndex(), '0x96eA4BbF71FEa3c9411C1Cefc555E9d7189695fA');
         await hotelProvider.__getContractInstance();
         throw new Error('should not have been called');
       } catch (e) {
@@ -42,7 +42,7 @@ describe('WTLibs.data-model.web3-uri.hotel', () => {
 
     it('should throw when we want hotel without an address', async () => {
       try {
-        const hotelProvider = await HotelDataProvider.createInstance(dataModel.web3Utils, dataModel.web3Contracts, await indexDataProvider.__getDeployedIndex());
+        const hotelProvider = await JsonHotelProvider.createInstance(dataModel.web3Utils, dataModel.web3Contracts, await indexDataProvider.__getDeployedIndex());
         await hotelProvider.__getContractInstance();
         throw new Error('should not have been called');
       } catch (e) {
@@ -52,7 +52,7 @@ describe('WTLibs.data-model.web3-uri.hotel', () => {
 
     it('should throw if we try to get data from network in a hotel without address', async () => {
       try {
-        const hotelProvider = await HotelDataProvider.createInstance(dataModel.web3Utils, dataModel.web3Contracts, await indexDataProvider.__getDeployedIndex());
+        const hotelProvider = await JsonHotelProvider.createInstance(dataModel.web3Utils, dataModel.web3Contracts, await indexDataProvider.__getDeployedIndex());
         await hotelProvider.ethBackedDataset.__syncRemoteData();
         throw new Error('should not have been called');
       } catch (e) {
@@ -62,7 +62,7 @@ describe('WTLibs.data-model.web3-uri.hotel', () => {
 
     it('should cache contract instances', async () => {
       assert.equal(getHotelContractSpy.callCount, 0);
-      const hotelProvider = await HotelDataProvider.createInstance(dataModel.web3Utils, dataModel.web3Contracts, await indexDataProvider.__getDeployedIndex(), '0x4a763f50dfe5cf4468b4171539e021a26fcee0cc');
+      const hotelProvider = await JsonHotelProvider.createInstance(dataModel.web3Utils, dataModel.web3Contracts, await indexDataProvider.__getDeployedIndex(), '0x4a763f50dfe5cf4468b4171539e021a26fcee0cc');
       assert.equal(getHotelContractSpy.callCount, 1);
       await hotelProvider.__getContractInstance();
       assert.equal(getHotelContractSpy.callCount, 1);
@@ -81,7 +81,7 @@ describe('WTLibs.data-model.web3-uri.hotel', () => {
         manager: '0xd39ca7d186a37bb6bf48ae8abfeb4c687dc8f906',
       });
 
-      const hotelProvider = await HotelDataProvider.createInstance(dataModel.web3Utils, dataModel.web3Contracts, await indexDataProvider.__getDeployedIndex(), result.address);
+      const hotelProvider = await JsonHotelProvider.createInstance(dataModel.web3Utils, dataModel.web3Contracts, await indexDataProvider.__getDeployedIndex(), result.address);
       const newName = 'Random changed name';
       hotelProvider.name = newName;
       await hotelProvider.updateOnNetwork(wallet, {
@@ -89,7 +89,7 @@ describe('WTLibs.data-model.web3-uri.hotel', () => {
         to: indexDataProvider.address,
       });
       assert.equal(await hotelProvider.name, newName);
-      let freshHotelProvider = await HotelDataProvider.createInstance(dataModel.web3Utils, dataModel.web3Contracts, await indexDataProvider.__getDeployedIndex(), result.address);
+      let freshHotelProvider = await JsonHotelProvider.createInstance(dataModel.web3Utils, dataModel.web3Contracts, await indexDataProvider.__getDeployedIndex(), result.address);
       assert.equal(await hotelProvider.name, await freshHotelProvider.name);
       // And remove the hotel to keep the data consistent
       await indexDataProvider.removeHotel(wallet, freshHotelProvider);
